API Reference

Solver and main API

Clarabel.Solver — Type

Solver{T <: AbstractFloat}()

Initializes an empty Clarabel solver that can be filled with problem data using:

setup!(solver, P, q, A, b, cones, [settings]).

Clarabel.setup! — Function

setup!(solver, P, q, A, b, cones, [settings])

Populates a Solver with a cost function defined by P and q, and one or more conic constraints defined by A, b and a description of a conic constraint composed of cones whose types and dimensions are specified by cones.

The solver will be configured to solve the following optimization problem:

min   1/2 x'Px + q'x
s.t.  Ax + s = b, s ∈ K

All data matrices must be sparse. The matrix P is assumed to be symmetric and positive semidefinite, and only the upper triangular part is used.

The cone K is a composite cone. To define the cone the user should provide a vector of cone specifications along with the appropriate dimensional information. For example, to generate a cone in the nonnegative orthant followed by a second order cone, use:

cones = [Clarabel.NonnegativeConeT(dim_1),
         Clarabel.SecondOrderConeT(dim_2)]

If the argument 'cones' is constructed incrementally, the should should initialize it as an empty array of the supertype for all allowable cones, e.g.

cones = Clarabel.SupportedCone[]
push!(cones,Clarabel.NonnegativeConeT(dim_1))
...

The optional argument settings can be used to pass custom solver settings:

settings = Clarabel.Settings(verbose = true)
setup!(model, P, q, A, b, cones, settings)

To solve the problem, you must make a subsequent call to solve!

Clarabel.solve! — Function

solve!(solver)

Computes the solution to the problem in a Clarabel.Solver previously defined in setup!.

Supported Cone Types

Clarabel.SupportedCone — Type

SupportedCone

An abstract type use by the Clarabel API used when passing cone specifications to the solver setup!. The currently supported concrete types are:

ZeroConeT : The zero cone. Used to define equalities.

NonnegativeConeT: The nonnegative orthant.

SecondOrderConeT: The second order / Lorentz / ice-cream cone.

PSDTriangleConeT: The positive semidefinite cone (triangular format).

Settings

Clarabel.Settings — Type

Clarabel.Settings{T}(kwargs) where {T <: AbstractFloat}

Creates a Clarabel Settings object that is used to pass user settings to the solver.

Argument	Default Value	Description

Main Algorithm Settings

max_iter	50	maximum number of iterations

time_limit	0	maximum run time (seconds)

verbose	true	verbose printing

tol_gap_abs	1e-8	absolute residual tolerance

tol_gap_rel	1e-8	relative residual tolerance

tol_feas	1e-5	feasibility check tolerance

tol\infeas\abs	1e-8	absolute infeasibility tolerance

tol\infeas\rel	1e-8	relative infeasibility tolerance

max_step_fraction	0.99	maximum interior point step length

Data Equilibration Settings

equilibrate_enable	true	enable data equilibration pre-scaling

equilibrate_max_iter	10	maximum equilibration scaling iterations

equilibrate_min_scaling	1e-4	minimum equilibration scaling allowed

equilibrate_max_scaling	1e+4	maximum equilibration scaling allowed

Linear Solver Settings

direct_kkt_solver	true	use a direct linear solver method (required true)

direct_solve_method	:qdldl	direct linear solver (:qdldl, :mkl or :cholmod)

static_regularization_enable	true	enable KKT static regularization

static_regularization_eps	1e-8	KKT static regularization parameter

dynamic_regularization_enable	true	enable KKT dynamic regularization

dynamic_regularization_eps	1e-13	KKT dynamic regularization threshold

dynamic_regularization_delta	2e-7	KKT dynamic regularization shift

iterative_refinement_enable	true	KKT solve with iterative refinement

iterative_refinement_reltol	1e-10	iterative refinement relative tolerance

iterative_refinement_abstol	1e-10	iterative refinement absolute tolerance

iterative_refinement_max_iter	10	iterative refinement maximum iterations

iterative_refinement_stop_ratio	2.0	iterative refinement stalling tolerance

Solver Status

Clarabel.SolverStatus — Type

SolverStatus

An Enum of of possible conditions set by solve!.

If no call has been made to solve!, then the SolverStatus is:

UNSOLVED: The algorithm has not started.

Otherwise:

SOLVED : Solver as terminated with a solution.

PRIMAL_INFEASIBLE : Problem is primal infeasible. Solution returned is a certificate of primal infeasibility.

DUAL_INFEASIBLE : Problem is dual infeasible. Solution returned is a certificate of dual infeasibility.

MAX_ITERATIONS : Iteration limit reached before solution or infeasibility certificate found.

MAX_TIME : Time limit reached before solution or infeasibility certificate found.